simplify date matching logic
Replace manual checking of each date element with using the alarm's
date string as a regex pattern to match against the current date/time.

The config file will now be created if it doesn't exist to avoid having
to check for its presence on each iteration of the main loop.

Additionally print a warning if a line's syntax is invalid and exec the
command with "sh -c" instead of eval to avoid modifying the current
environment.